refactor(web): extract the files panel's column preference and debounced search into hooks
WorkspaceFilesPanel.tsx sat at 995 lines with its state and handlers
undivided. The two cleanly separable concerns move out, verbatim in
behavior and mostly verbatim in text:

- use-browser-columns.ts owns the BrowserColumns type, the guarded
  localStorage read/write (same key, whiteboard.document-browser.columns.v1),
  and the choose handler.
- use-debounced-document-search.ts owns the query/hits/degraded state and
  the debounced, stale-marking source search, plus the resetResults the
  panel's workspace-switch effect calls.

995 -> 908 lines. What remains is the panel's list/selection/menu/rename/
pin orchestration, which all writes the same documents+selected state —
splitting that would trade one coherent owner for prop-bag plumbing, so it
deliberately stays. The rendering was already decomposed into the sibling
components (WorkspaceFolderTree, FolderContentsList, SearchResults, ...).

Behavior-preserving: the workspace-files jsdom suite (28 files, 205 tests)
and browser suites (7 tests) ran un-weakened, and the real built app was
driven in Chromium — column choice persists across reload, the debounced
search surfaces a created note, clearing the query restores the browser.

Resolves the audit finding workspace-files-panel-split.
